+++ cli/ballerina-packerina/src/test/resources/compile_command_test_project/src/mymodule/main.bal
@@ -1,5 +1,5 @@
-import testOrg/myimport;
+// import testOrg/myimport;
 
 public function main(string... args) {
-    int x = myimport:foo();
+    // int x = myimport:foo();
 }
