+++ misc/testerina/modules/testerina-core/src/main/java/org/ballerinalang/testerina/core/entity/TesterinaFunction.java
@@ -17,20 +17,20 @@
  */
 package org.ballerinalang.testerina.core.entity;
 
-import org.ballerinalang.bre.bvm.BVMExecutor;
+import org.ballerinalang.jvm.scheduling.Scheduler;
+import org.ballerinalang.jvm.scheduling.Strand;
+import org.ballerinalang.jvm.values.ErrorValue;
+import org.ballerinalang.jvm.values.FutureValue;
 import org.ballerinalang.model.values.BValue;
-import org.ballerinalang.testerina.core.TesterinaRegistry;
-import org.ballerinalang.testerina.util.TesterinaUtils;
-import org.ballerinalang.util.LaunchListener;
-import org.ballerinalang.util.codegen.FunctionInfo;
-import org.ballerinalang.util.codegen.PackageInfo;
-import org.ballerinalang.util.codegen.ProgramFile;
-import org.ballerinalang.util.debugger.Debugger;
 import org.ballerinalang.util.exceptions.BallerinaException;
+import org.wso2.ballerinalang.compiler.tree.BLangFunction;
+import org.wso2.ballerinalang.compiler.tree.BLangIdentifier;
 
+import java.lang.reflect.InvocationTargetException;
+import java.lang.reflect.Method;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.ServiceLoader;
+import java.util.function.Function;
 
 /**
  * TesterinaFunction entity class.
@@ -40,12 +40,12 @@ public class TesterinaFunction {
     private String name;
     private Type type;
 
-    public FunctionInfo getbFunction() {
+    public BLangFunction getbFunction() {
         return bFunction;
     }
 
-    private FunctionInfo bFunction;
-    private ProgramFile programFile;
+    private BLangFunction bFunction;
+    private Class<?> programFile;
     private boolean runTest = true;
 
     // Annotation info
@@ -77,34 +77,17 @@ public class TesterinaFunction {
         }
     }
 
-    public TesterinaFunction(ProgramFile programFile, FunctionInfo bFunction, Type type) {
-        this.name = bFunction.getName();
+    public TesterinaFunction(Class<?> programFile, BLangFunction bFunction, Type type) {
+        this.name = bFunction.getName().getValue();
         this.type = type;
         this.bFunction = bFunction;
         this.programFile = programFile;
     }
 
     public BValue[] invoke() throws BallerinaException {
-        if (this.type == Type.TEST_INIT) {
-            // Load launcher listeners
-            ServiceLoader<LaunchListener> listeners = ServiceLoader.load(LaunchListener.class);
-            listeners.forEach(listener -> listener.beforeRunProgram(true));
-
-            // Invoke init functions
-            TesterinaUtils.invokePackageInitFunctions(programFile);
-            TesterinaUtils.invokePackageTestInitFunctions(programFile);
-
-            //  Invoke start functions
-            TesterinaUtils.invokePackageStartFunctions(programFile);
-            TesterinaUtils.invokePackageTestStartFunctions(programFile);
-
-            listeners.forEach(listener -> listener.afterRunProgram(true));
-
-            TesterinaRegistry.getInstance().addInitializedPackage(programFile.getEntryPkgName());
-            return new BValue[]{};
-        } else {
-            return invoke(new BValue[]{});
-        }
+        final Scheduler scheduler = new Scheduler(4, false);
+        runOnSchedule(programFile, bFunction.name, scheduler);
+        return new BValue[0];
     }
 
     /**
@@ -113,7 +96,7 @@ public class TesterinaFunction {
      * @throws BallerinaException exception is thrown
      */
     public void invokeStopFunctions() throws BallerinaException {
-        for (PackageInfo info : programFile.getPackageInfoEntries()) {
+        /*for (PackageInfo info : programFile.getPackageInfoEntries()) {
             BVMExecutor.executeFunction(programFile, info.getStopFunctionInfo());
         }
 
@@ -121,7 +104,7 @@ public class TesterinaFunction {
             if (info.getTestStopFunctionInfo() != null) {
                 BVMExecutor.executeFunction(programFile, info.getTestStopFunctionInfo());
             }
-        }
+        }*/
     }
 
     /**
@@ -131,7 +114,8 @@ public class TesterinaFunction {
      * @return a BValue array
      */
     public BValue[] invoke(BValue[] args) {
-        return BVMExecutor.executeFunction(programFile, bFunction, args);
+        // return BVMExecutor.executeFunction(programFile, bFunction, args);
+        return new BValue[0];
     }
 
     public String getName() {
@@ -166,11 +150,47 @@ public class TesterinaFunction {
         this.runTest = false;
     }
 
-    private static void initDebugger(ProgramFile programFile, Debugger debugger) {
-        programFile.setDebugger(debugger);
-        if (debugger.isDebugEnabled()) {
-            debugger.init();
-            debugger.waitTillDebuggeeResponds();
+
+
+    private static void runOnSchedule(Class<?> initClazz, BLangIdentifier name, Scheduler scheduler1) {
+        String funcName = cleanupFunctionName(name);
+        try {
+            final Method method = initClazz.getDeclaredMethod(funcName, Strand.class);
+            Scheduler scheduler = scheduler1;
+            //TODO fix following method invoke to scheduler.schedule()
+            Function<Object[], Object> func = objects -> {
+                try {
+                    return method.invoke(null, objects[0]);
+                } catch (InvocationTargetException e) {
+                    e.printStackTrace();
+                    throw (RuntimeException) e.getTargetException();
+                } catch (IllegalAccessException e) {
+                    throw new BallerinaException("Method has private access", e);
+                }
+            };
+            final FutureValue out = scheduler.schedule(new Object[1], func, null, null, null);
+            scheduler.start();
+            final Throwable t = out.panic;
+            if (t != null) {
+                if (t instanceof org.ballerinalang.jvm.util.exceptions.BLangRuntimeException) {
+                    throw new org.ballerinalang.util.exceptions.BLangRuntimeException(t.getMessage());
+                }
+                if (t instanceof org.ballerinalang.jvm.util.exceptions.BallerinaConnectorException) {
+                    throw new org.ballerinalang.util.exceptions.BLangRuntimeException(t.getMessage());
+                }
+                if (t instanceof ErrorValue) {
+                    throw new org.ballerinalang.util.exceptions.BLangRuntimeException(
+                            "error: " + ((ErrorValue) t).getPrintableStackTrace());
+                }
+                throw (RuntimeException) t;
+            }
+        } catch (NoSuchMethodException e) {
+            throw new RuntimeException("Error while invoking function '" + funcName + "'", e);
         }
     }
+
+    private static String cleanupFunctionName(BLangIdentifier name) {
+        return name.value.replaceAll("[.:/<>]", "_");
+    }
+
 }
